requirements.txt
aiogram==3.13.1
python-dotenv
opencv-python-headless
aiofiles
pypotrace
Pillow
numpy